Move ip.c to tcp.c
