+++ src/error.c
@@ -37,16 +37,16 @@ static const ERROR_NAME global_error_table[] = {
 	{RC_HTTPS_NO_TRUSTED_CA_STORE, "RC_HTTPS_NO_TRUSTED_CA_STORE"},
 	{RC_PIDFILE_EXISTS_ALREADY, "RC_PIDFILE_EXISTS_ALREADY"},
 
-	{RC_IP_SOCKET_CREATE_ERROR, "RC_IP_SOCKET_CREATE_ERROR"},
-	{RC_IP_BAD_PARAMETER, "RC_IP_BAD_PARAMETER"},
-	{RC_IP_INVALID_REMOTE_ADDR, "RC_IP_INVALID_REMOTE_ADDR"},
-	{RC_IP_CONNECT_FAILED, "RC_IP_CONNECT_FAILED"},
-	{RC_IP_SEND_ERROR, "RC_IP_SEND_ERROR"},
-	{RC_IP_RECV_ERROR, "RC_IP_RECV_ERROR"},
-	{RC_IP_OBJECT_NOT_INITIALIZED, "RC_IP_OBJECT_NOT_INITIALIZED"},
-	{RC_IP_OS_SOCKET_INIT_FAILED, "RC_IP_OS_SOCKET_INIT_FAILED"},
-
+	{RC_TCP_SOCKET_CREATE_ERROR, "RC_TCP_SOCKET_CREATE_ERROR"},
+	{RC_TCP_BAD_PARAMETER, "RC_TCP_BAD_PARAMETER"},
+	{RC_TCP_INVALID_REMOTE_ADDR, "RC_TCP_INVALID_REMOTE_ADDR"},
+	{RC_TCP_CONNECT_FAILED, "RC_TCP_CONNECT_FAILED"},
+	{RC_TCP_SEND_ERROR, "RC_TCP_SEND_ERROR"},
+	{RC_TCP_RECV_ERROR, "RC_TCP_RECV_ERROR"},
 	{RC_TCP_OBJECT_NOT_INITIALIZED, "RC_TCP_OBJECT_NOT_INITIALIZED"},
+	{RC_TCP_OS_SOCKET_INIT_FAILED, "RC_TCP_OS_SOCKET_INIT_FAILED"},
+
+//	{RC_TCP_OBJECT_NOT_INITIALIZED, "RC_TCP_OBJECT_NOT_INITIALIZED"},
 
 	{RC_HTTP_OBJECT_NOT_INITIALIZED, "RC_HTTP_OBJECT_NOT_INITIALIZED"},
 	{RC_HTTP_BAD_PARAMETER, "RC_HTTP_BAD_PARAMETER"},
